Introduce navigation component
... to better and safely handle fragments:

MainActivity code has been fragmented into new fragments:
1. LoadingFragment: the data is loaded here after the user has
granted permissions to access MediaStore
2. MainFragmet: this is where the data is presented
3. Player fragment: media player

Navigation is ensured thanks to Navigation Component
