+++ project/app/src/main/java/com/iven/musicplayergo/adapters/LovedSongsAdapter.kt
@@ -15,21 +15,20 @@ import com.iven.musicplayergo.helpers.DialogHelper
 import com.iven.musicplayergo.helpers.MusicOrgHelper
 import com.iven.musicplayergo.models.SavedMusic
 import com.iven.musicplayergo.player.MediaPlayerHolder
-import com.iven.musicplayergo.ui.UIControlInterface
 
 class LovedSongsAdapter(
     private val context: Context,
-    private val lovedSongsDialog: MaterialDialog,
-    private val uiControlInterface: UIControlInterface
+    private val lovedSongsDialog: MaterialDialog
 ) :
     RecyclerView.Adapter<LovedSongsAdapter.LoveHolder>() {
 
     private var mLovedSongs = goPreferences.lovedSongs?.toMutableList()
 
+    private val mMediaPlayerInterface = MediaPlayerHolder.getInstance().mediaPlayerInterface
     fun swapSongs(lovedSongs: MutableList<SavedMusic>?) {
         mLovedSongs = lovedSongs
         notifyDataSetChanged()
-        uiControlInterface.onLovedSongsUpdate(false)
+        mMediaPlayerInterface?.onLovedSongUpdate(false)
         if (mLovedSongs?.isEmpty()!!) {
             lovedSongsDialog.dismiss()
         }
@@ -75,11 +74,12 @@ class LovedSongsAdapter(
 
             itemView.apply {
                 setOnClickListener {
-                    MediaPlayerHolder.getInstance().isSongFromLovedSongs =
+                    val mediaPlayerHolder = MediaPlayerHolder.getInstance()
+                    mediaPlayerHolder.isSongFromLovedSongs =
                         Pair(true, lovedSong?.startFrom!!)
                     MusicOrgHelper.getSongForRestore(lovedSong)
                         .apply {
-                            uiControlInterface.onSongSelected(
+                            mediaPlayerHolder.mediaPlayerInterface?.onSongSelected(
                                 this,
                                 MusicOrgHelper.getAlbumSongs(
                                     artist,
