+++ project/app/src/main/java/com/iven/musicplayergo/extensions/ViewExts.kt
@@ -1,31 +1,27 @@
 package com.iven.musicplayergo.extensions
 
-import android.animation.Animator
-import android.animation.ValueAnimator
 import android.content.Context
+import android.graphics.Color
 import android.text.Html
 import android.text.SpannableString
 import android.text.Spanned
 import android.text.style.ForegroundColorSpan
 import android.view.MenuItem
 import android.view.View
-import android.view.ViewAnimationUtils
 import android.view.ViewTreeObserver
+import android.view.Window
 import android.widget.Toast
 import androidx.annotation.ColorInt
-import androidx.core.animation.doOnEnd
 import androidx.core.content.ContextCompat
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.interpolator.view.animation.FastOutSlowInInterpolator
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearSmoothScroller
 import androidx.recyclerview.widget.RecyclerView
-import com.google.android.material.animation.ArgbEvaluatorCompat
-import com.iven.musicplayergo.R
-import com.iven.musicplayergo.helpers.ThemeHelper
 import com.iven.musicplayergo.helpers.VersioningHelper
-import kotlin.math.max
+
+fun Window.handleTransparentSystemBars() {
+    statusBarColor = Color.TRANSPARENT
+    navigationBarColor = Color.TRANSPARENT
+}
 
 // viewTreeObserver extension to measure layout params
 // https://antonioleiva.com/kotlin-ongloballayoutlistener/
@@ -57,74 +53,6 @@ fun MenuItem.setTitleColor(color: Int) {
         setSpan(ForegroundColorSpan(color), 0, length, SpannableString.SPAN_EXCLUSIVE_EXCLUSIVE)
         title = this
     }
-}
-
-fun FragmentManager.addFragment(fragment: Fragment, tag: String?) {
-    beginTransaction().apply {
-        addToBackStack(null)
-        add(
-            R.id.container,
-            fragment,
-            tag
-        )
-        commit()
-    }
-}
-
-fun FragmentManager.isFragment(fragmentTag: String): Boolean {
-    val df = findFragmentByTag(fragmentTag)
-    return df != null && df.isVisible && df.isAdded
-}
-
-fun View.createCircularReveal(isErrorFragment: Boolean, show: Boolean): Animator {
-
-    val revealDuration: Long = if (isErrorFragment) 1500 else 500
-    val radius = max(width, height).toFloat()
-
-    val startRadius = if (show) 0f else radius
-    val finalRadius = if (show) radius else 0f
-
-    val cx = if (isErrorFragment) width / 2 else 0
-    val cy = if (isErrorFragment) height / 2 else 0
-    val animator =
-        ViewAnimationUtils.createCircularReveal(
-            this,
-            cx,
-            cy,
-            startRadius,
-            finalRadius
-        ).apply {
-            interpolator = FastOutSlowInInterpolator()
-            duration = revealDuration
-            doOnEnd {
-                if (!show) handleViewVisibility(false)
-            }
-            start()
-        }
-
-    val windowBackground = R.color.windowBackground.decodeColor(context)
-    val red = R.color.red.decodeColor(context)
-    val accent = if (!show) windowBackground else ThemeHelper.resolveThemeAccent(context)
-
-    val startColor = if (isErrorFragment) red else accent
-    val endColor = if (show) windowBackground else red
-
-    ValueAnimator().apply {
-        setIntValues(startColor, endColor)
-        setEvaluator(ArgbEvaluatorCompat())
-        addUpdateListener { valueAnimator -> setBackgroundColor((valueAnimator.animatedValue as Int)) }
-        duration = revealDuration
-        if (isErrorFragment) doOnEnd {
-            background =
-                ThemeHelper.createColouredRipple(
-                    context,
-                    R.color.red.decodeColor(context),
-                    R.drawable.ripple
-                )
-        }
-        start()
-    }
-    return animator
 }
 
 // https://stackoverflow.com/a/53986874
