+++ project/app/src/main/java/com/iven/musicplayergo/navigation/ErrorFragment.kt
@@ -1,28 +1,20 @@
-package com.iven.musicplayergo.fragments
+package com.iven.musicplayergo.navigation
 
 import android.content.Context
 import android.os.Bundle
 import android.view.View
-import androidx.core.animation.doOnEnd
 import androidx.fragment.app.Fragment
 import com.iven.musicplayergo.GoConstants
 import com.iven.musicplayergo.R
 import com.iven.musicplayergo.databinding.FragmentErrorBinding
-import com.iven.musicplayergo.extensions.afterMeasured
-import com.iven.musicplayergo.extensions.createCircularReveal
 import com.iven.musicplayergo.extensions.decodeColor
 import com.iven.musicplayergo.goPreferences
-import com.iven.musicplayergo.ui.UIControlInterface
+import com.iven.musicplayergo.interfaces.UIControlInterface
 
-/**
- * A simple [Fragment] subclass.
- * Use the [ErrorFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
 class ErrorFragment : Fragment(R.layout.fragment_error) {
 
     private lateinit var mErrorFragmentBinding: FragmentErrorBinding
-    private lateinit var mUIControlInterface: UIControlInterface
+    private var mUIControlInterface: UIControlInterface? = null
 
     private var mErrorString = R.string.perm_rationale
     private var mErrorIcon = R.drawable.ic_folder
@@ -30,7 +22,7 @@ class ErrorFragment : Fragment(R.layout.fragment_error) {
     override fun onAttach(context: Context) {
         super.onAttach(context)
 
-        arguments?.getString(TAG_ERROR)?.let { errorType ->
+        arguments?.getString(GoConstants.TAG_ERROR)?.let { errorType ->
 
             when (errorType) {
                 GoConstants.TAG_NO_MUSIC -> {
@@ -61,41 +53,21 @@ class ErrorFragment : Fragment(R.layout.fragment_error) {
         super.onViewCreated(view, savedInstanceState)
 
         mErrorFragmentBinding = FragmentErrorBinding.bind(view)
-        mErrorFragmentBinding.errorMessage.text = getString(mErrorString)
-        mErrorFragmentBinding.errorIcon.setImageResource(mErrorIcon)
-        mErrorFragmentBinding.root.setOnClickListener { mUIControlInterface.onCloseActivity() }
 
-        mErrorFragmentBinding.errorToolbar.setNavigationOnClickListener {
-            mUIControlInterface.onCloseActivity()
-        }
-
-        mErrorFragmentBinding.root.afterMeasured {
-            createCircularReveal(isErrorFragment = true, show = true).doOnEnd {
-                if (!goPreferences.isEdgeToEdge) {
-                    requireActivity().window.apply {
-                        val red = R.color.red.decodeColor(requireContext())
-                        statusBarColor = red
-                        navigationBarColor = red
-                    }
-                }
+        mErrorFragmentBinding.apply {
+            errorMessage.text = getString(mErrorString)
+            errorIcon.setImageResource(mErrorIcon)
+            root.setOnClickListener { requireActivity().recreate() }
+            errorToolbar.setNavigationOnClickListener {
+                mUIControlInterface?.onCloseActivity()
             }
         }
-    }
 
-    companion object {
-
-        private const val TAG_ERROR = "WE_HAVE_A_PROBLEM_HOUSTON"
-
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @return A new instance of fragment ErrorFragment.
-         */
-        @JvmStatic
-        fun newInstance(errorType: String) = ErrorFragment().apply {
-            arguments = Bundle().apply {
-                putString(TAG_ERROR, errorType)
+        if (!goPreferences.isEdgeToEdge) {
+            requireActivity().window.apply {
+                val red = R.color.red.decodeColor(requireContext())
+                statusBarColor = red
+                navigationBarColor = red
             }
         }
     }
