--- project/app/src/main/java/com/iven/musicplayergo/player/MediaPlayerInterface.kt
@@ -1,13 +1,0 @@
-package com.iven.musicplayergo.player
-
-interface MediaPlayerInterface {
-    fun onPositionChanged(position: Int)
-    fun onStateChanged()
-    fun onPlaybackCompleted()
-    fun onClose()
-    fun onUpdateRepeatStatus()
-    fun onQueueEnabled()
-    fun onQueueCleared()
-    fun onQueueStartedOrEnded(started: Boolean)
-    fun onSaveSong()
-}
