+++ project/app/src/main/res/layout-land/now_playing.xml
@@ -5,6 +5,7 @@
     android:layout_height="wrap_content"
     android:clickable="false"
     android:focusable="false"
+    android:fitsSystemWindows="true"
     android:orientation="vertical">
 
     <LinearLayout
@@ -19,7 +20,6 @@
             android:layout_marginBottom="8dp"
             android:layout_weight="1"
             android:background="@drawable/ripple"
-            android:onClick="openPlayingArtistAlbum"
             android:orientation="vertical"
             android:paddingTop="8dp"
             android:paddingBottom="8dp"
