+++ project/app/src/main/res/layout-land/now_playing_controls.xml
@@ -19,7 +19,6 @@
             android:layout_gravity="center"
             android:layout_weight="1"
             android:background="?android:attr/actionBarItemBackground"
-            android:onClick="openEqualizer"
             app:srcCompat="@drawable/ic_equalizer"
             tools:ignore="ContentDescription" />
