+++ project/app/src/main/res/layout/fragment_loading_fragment.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ProgressBar android:id="@+id/loading_progress_bar"
+    android:layout_width="@android:dimen/app_icon_size"
+    android:layout_height="@android:dimen/app_icon_size"
+    android:layout_gravity="center"
+    android:clickable="false"
+    android:fitsSystemWindows="true"
+    android:indeterminateTint="?attr/colorPrimary"
+    android:indeterminateTintMode="src_in"
+    xmlns:android="http://schemas.android.com/apk/res/android" />
