+++ project/app/src/main/res/layout/fragment_main.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:fitsSystemWindows="true"
+    android:orientation="vertical"
+    tools:context=".ui.MainActivity">
+
+    <androidx.viewpager2.widget.ViewPager2
+        android:id="@+id/view_pager2"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1" />
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:orientation="vertical"
+        android:layout_height="wrap_content">
+
+        <!-- CONTROLS PANEL -->
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/fragment_player"
+            android:name="com.iven.musicplayergo.player.PlayerFragment"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
+
+        <!-- TAB LAYOUT -->
+        <com.google.android.material.tabs.TabLayout
+            android:id="@+id/tab_layout"
+            android:layout_width="match_parent"
+            android:layout_height="32dp"
+            app:tabGravity="fill"
+            android:background="@color/windowBackground"
+            app:tabIndicator="@drawable/tab_indicator"
+            app:tabMode="fixed"
+            app:tabRippleColor="?attr/colorPrimary" />
+    </LinearLayout>
+
+</LinearLayout>
+
+
+
