+++ project/app/src/main/res/layout/fragment_settings.xml
@@ -3,6 +3,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:viewBindingIgnore="true"
     android:orientation="vertical">
 
     <androidx.appcompat.widget.Toolbar
