+++ project/app/src/main/res/layout/now_playing.xml
@@ -5,6 +5,7 @@
     android:layout_height="wrap_content"
     android:clickable="false"
     android:focusable="false"
+    android:fitsSystemWindows="true"
     android:orientation="vertical">
 
     <LinearLayout
@@ -12,7 +13,6 @@
         android:layout_height="wrap_content"
         android:layout_marginBottom="4dp"
         android:background="@drawable/ripple"
-        android:onClick="openPlayingArtistAlbum"
         android:orientation="vertical"
         android:paddingTop="16dp"
         android:paddingBottom="16dp"
