--- project/app/src/main/res/layout/player_controls_panel.xml
@@ -1,118 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<merge xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="@layout/main_activity">
-
-    <LinearLayout
-        android:id="@+id/player_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:alpha="0"
-        android:background="@color/windowBackground"
-        android:elevation="@dimen/search_bar_elevation"
-        android:orientation="vertical"
-        tools:context=".ui.MainActivity">
-
-        <ProgressBar
-            android:id="@+id/song_progress"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="match_parent"
-            android:layout_height="2dp"
-            android:clickable="false"
-            android:indeterminateTint="?attr/colorPrimary"
-            android:indeterminateTintMode="src_in"
-            android:progressBackgroundTint="?attr/colorControlNormal" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:baselineAligned="false"
-            android:orientation="horizontal"
-            android:padding="4dp">
-
-            <LinearLayout
-                android:id="@+id/playing_songs_container"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_weight="1"
-                android:background="@drawable/ripple_alt"
-                android:onClick="openNowPlaying"
-                android:orientation="vertical"
-                android:paddingStart="@dimen/player_controls_padding_start"
-                android:paddingTop="@dimen/player_controls_padding_normal"
-                android:paddingEnd="@dimen/player_controls_padding_end"
-                android:paddingBottom="@dimen/player_controls_padding_normal">
-
-                <TextView
-                    android:id="@+id/playing_song"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:ellipsize="marquee"
-                    android:fontFamily="@font/source_sans_pro_black"
-                    android:marqueeRepeatLimit="marquee_forever"
-                    android:singleLine="true"
-                    android:text="@string/app_name"
-                    android:textAlignment="textStart"
-                    android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
-                    android:textColor="?android:attr/textColorPrimary"
-                    android:textStyle="bold" />
-
-                <TextView
-                    android:id="@+id/playing_artist"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:ellipsize="marquee"
-                    android:marqueeRepeatLimit="marquee_forever"
-                    android:singleLine="true"
-                    android:text="@string/versionName"
-                    android:textAlignment="textStart"
-                    android:textAppearance="@style/TextAppearance.MaterialComponents.Caption"
-                    android:textColor="?android:attr/textColorPrimary" />
-            </LinearLayout>
-
-            <ImageButton
-                android:id="@+id/play_pause_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:background="?android:attr/actionBarItemBackground"
-                android:padding="8dp"
-                app:srcCompat="@drawable/ic_play"
-                tools:ignore="ContentDescription" />
-
-            <ImageButton
-                android:id="@+id/queue_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:background="?android:attr/actionBarItemBackground"
-                android:onClick="openQueueDialog"
-                android:padding="8dp"
-                app:srcCompat="@drawable/ic_queue_music"
-                tools:ignore="ContentDescription" />
-
-            <ImageButton
-                android:id="@+id/loved_songs_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:background="?android:attr/actionBarItemBackground"
-                android:onClick="openLovedSongsDialog"
-                android:padding="8dp"
-                app:srcCompat="@drawable/ic_favorite"
-                tools:ignore="ContentDescription" />
-        </LinearLayout>
-
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/tab_layout"
-            android:layout_width="match_parent"
-            android:layout_height="32dp"
-            android:background="@color/windowBackground"
-            app:tabGravity="fill"
-            app:tabIndicator="@drawable/tab_indicator"
-            app:tabMode="fixed"
-            app:tabRippleColor="?attr/colorPrimary" />
-    </LinearLayout>
-</merge>
